Introduce the `remoting` image

+++ container-images/scripts/build_llama_and_whisper.sh
@@ -18,6 +18,12 @@ available() {
   command -v "$1" >/dev/null
 }
 
+dnf_install_remoting() {
+    LIBDRM_VERSION=2.4.123-2.el9.aarch64
+    curl -Ssf https://mirror.stream.centos.org/9-stream/AppStream/aarch64/os/Packages/libdrm-devel-${LIBDRM_VERSION}.rpm -O
+    rpm -i --nosignature --nodeps libdrm-devel-${LIBDRM_VERSION}.rpm
+}
+
 dnf_install_intel_gpu() {
   local intel_rpms=("intel-oneapi-mkl-sycl-devel" "intel-oneapi-dnnl-devel"
     "intel-oneapi-compiler-dpcpp-cpp" "intel-level-zero"
@@ -164,9 +170,16 @@ dnf_install() {
     dnf_install_intel_gpu
   elif [ "$containerfile" = "cann" ]; then
     dnf_install_cann
+  elif [ "$containerfile" = "remoting" ]; then
+    dnf_install_remoting
   fi
 
   dnf_install_ffmpeg
+
+  if [[ "${RAMALAMA_IMAGE_INCLUDE_DEBUG:-}" == y ]]; then
+      dnf install -y gdb strace
+  fi
+
   dnf -y clean all
 }
 
@@ -242,41 +255,53 @@ configure_common_flags() {
   musa)
     common_flags+=("-DGGML_MUSA=ON")
     ;;
+  remoting)
+    common_flags+=("-DGGML_REMOTINGFRONTEND=ON")
+    ;;
   esac
 }
 
 clone_and_build_whisper_cpp() {
   local whisper_flags=("${common_flags[@]}")
   # last time we tried to upgrade the whisper sha, rocm build broke
-  local whisper_cpp_sha="d682e150908e10caa4c15883c633d7902d385237"
+  local whisper_cpp_sha="${WHISPER_CPP_PULL_REF:-main}"
   whisper_flags+=("-DBUILD_SHARED_LIBS=OFF")
   # See: https://github.com/ggml-org/llama.cpp/blob/master/docs/build.md#compilation-options
   if [ "$containerfile" = "musa" ]; then
     whisper_flags+=("-DCMAKE_POSITION_INDEPENDENT_CODE=ON")
   fi
 
-  git clone https://github.com/ggerganov/whisper.cpp
+  git clone https://github.com/kpouget/whisper.cpp
   cd whisper.cpp
   git submodule update --init --recursive
-  git reset --hard "$whisper_cpp_sha"
+  git fetch origin "$whisper_cpp_sha"
+  git reset --hard FETCH_HEAD
   cmake_steps "${whisper_flags[@]}"
   mkdir -p "$install_prefix/bin"
   cd ..
-  rm -rf whisper.cpp
+  if [[ "${RAMALAMA_IMAGE_INCLUDE_DEBUG:-}" != y ]]; then
+      rm -rf whisper.cpp
+  fi
 }
 
 clone_and_build_llama_cpp() {
-  local llama_cpp_sha="97340b4c9924be86704dbf155e97c8319849ee19"
+  local containerfile=$1
+  local llama_cpp_sha="${LLAMA_CPP_PULL_REF:-main}"
   local install_prefix
   install_prefix=$(set_install_prefix)
-  git clone https://github.com/ggml-org/llama.cpp
+  git clone https://github.com/kpouget/llama.cpp
   cd llama.cpp
   git submodule update --init --recursive
-  git reset --hard "$llama_cpp_sha"
+  git fetch origin "$llama_cpp_sha"
+  git reset --hard FETCH_HEAD
   cmake_steps "${common_flags[@]}"
-  install -m 755 build/bin/rpc-server "$install_prefix"/bin/rpc-server
+  if [[ "$containerfile" != "remoting" ]]; then
+      install -m 755 build/bin/rpc-server "$install_prefix"/bin/rpc-server
+  fi
   cd ..
-  rm -rf llama.cpp
+  if [[ "${RAMALAMA_IMAGE_INCLUDE_DEBUG:-}" != y ]]; then
+      rm -rf llama.cpp
+  fi
 }
 
 install_ramalama() {
@@ -318,7 +343,10 @@ main() {
   if [ "$uname_m" != "s390x" ]; then
     clone_and_build_whisper_cpp
   fi
-  common_flags+=("-DLLAMA_CURL=ON" "-DGGML_RPC=ON")
+  common_flags+=("-DLLAMA_CURL=ON")
+  if [[ "$containerfile" != "remoting" ]]; then
+      common_flags+=("-DGGML_RPC=ON")
+  fi
   case "$containerfile" in
   ramalama)
     if [ "$uname_m" = "x86_64" ] || [ "$uname_m" = "aarch64" ]; then
@@ -329,7 +357,7 @@ main() {
     ;;
   esac
 
-  clone_and_build_llama_cpp
+  clone_and_build_llama_cpp "$containerfile"
   available dnf && dnf_remove
   rm -rf /var/cache/*dnf* /opt/rocm-*/lib/*/library/*gfx9*
   ldconfig # needed for libraries
